LayoutTests/platform/ios-sim-deprecated/transforms/3d/hit-testing/rotated-hit-test2-expected.txt

CONSOLE MESSAGE: line 15: TypeError: null is not an object (evaluating 'hit.id')
Element at 170, 250 has id "left": PASS
Element at 260, 170 has id "top": PASS
